feat(admin-ui): remove admin-ui-react from bundle
