--- packages/admin-ui-react/README.md
@@ -1,3 +1,0 @@
-# @vtex/admin-ui-react
-
-This package binds admin-ui-core to react, for ease of use.
