--- packages/admin-ui-react/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "@vtex/admin-ui-react",
-  "license": "MIT",
-  "version": "0.10.1-dev.14",
-  "main": "dist/vtex-admin-ui-react.cjs.js",
-  "module": "dist/vtex-admin-ui-react.esm.js",
-  "types": "dist/vtex-admin-ui-react.cjs.d.ts",
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org"
-  },
-  "files": [
-    "dist",
-    "src"
-  ],
-  "engines": {
-    "node": ">=10"
-  },
-  "scripts": {
-    "prebuild": "rm -rf dist",
-    "dev": "rollup -c -w",
-    "build": "npm run prebuild && rollup -c"
-  },
-  "repository": {
-    "directory": "packages/react",
-    "type": "git",
-    "url": "git+https://github.com/vtex/admin-ui.git"
-  },
-  "bugs": {
-    "url": "https://github.com/vtex/admin-ui/issues"
-  },
-  "peerDependencies": {
-    "@vtex/admin-ui-core": ">=0.x",
-    "@vtex/admin-ui-util": ">=0.x",
-    "@vtex/phosphor-icons": ">=0.x",
-    "react": "^16.9 || ^17.0",
-    "react-dom": "^16.9 || ^17.0"
-  },
-  "devDependencies": {
-    "@types/react-helmet": "^6.1.5",
-    "@vtex/admin-ui-core": "^0.13.1-dev.14",
-    "@vtex/admin-ui-util": "^0.5.1-dev.13",
-    "@vtex/phosphor-icons": "0.2.8"
-  },
-  "dependencies": {
-    "focus-visible": "^5.2.0",
-    "react-helmet": "^6.1.0",
-    "react-is": "^17.0.2",
-    "tiny-invariant": "^1.1.0",
-    "tiny-warning": "^1.0.3"
-  }
-}
