--- packages/admin-ui-react/rollup.config.js
@@ -1,3 +1,0 @@
-import rollupConfig from '../../rollup.config'
-
-export default rollupConfig
