--- packages/admin-ui-react/src/index.ts
@@ -1,2 +1,0 @@
-export * from './hooks'
-export * from './system'
