--- packages/admin-ui-react/src/system/index.ts
@@ -1,5 +1,0 @@
-export * from './context'
-export * from './create-system'
-export * from './icons'
-export * from './types'
-export * from './system'
