--- packages/admin-ui-react/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "declarationDir": "declarations"
-  },
-  "include": ["./src", "./typings/vtex.phosphor.d.ts"],
-  "exclude": [
-    "node_modules",
-    "dist",
-    "**/*.test.*",
-    "**/*.stories.*",
-    "**/*test-utils*"
-  ]
-}
