+++ packages/admin-ui/package.json
@@ -40,9 +40,11 @@
     "intersection-observer": "^0.12.2",
     "mutationobserver-shim": "^0.3.7",
     "react-beautiful-dnd": "^13.0.0",
-    "react-window": "^1.8.5"
+    "react-window": "^1.8.5",
+    "@types/react-helmet": "6.1.6"
   },
   "dependencies": {
+    "focus-visible": "^5.2.0",
     "@babel/core": "^7.17.10",
     "@react-aria/i18n": "^3.3.5",
     "@react-aria/listbox": "^3.4.1",
@@ -52,7 +54,6 @@
     "@react-stately/list": "^3.4.1",
     "@vtex/admin-ui-core": "^0.13.1-dev.14",
     "@vtex/admin-ui-hooks": "^0.5.1-dev.13",
-    "@vtex/admin-ui-react": "^0.10.1-dev.14",
     "@vtex/admin-ui-util": "^0.5.1-dev.13",
     "@vtex/phosphor-icons": "0.2.8",
     "ariakit": "2.0.0-next.43",
@@ -62,6 +63,7 @@
     "jotai": "^1.9.2",
     "react-collapsed": "^3.3.0",
     "react-device-detect": "^1.15.0",
+    "react-helmet": "6.1.0",
     "react-is": "^17.0.1",
     "reakit": "^1.3.11",
     "tiny-invariant": "^1.1.0",
