+++ packages/admin-ui/src/columns/columns.tsx
@@ -1,10 +1,10 @@
 import type { ComponentPropsWithRef, Ref } from 'react'
 import React, { forwardRef } from 'react'
-import type { ResponsiveProp } from '@vtex/admin-ui-react'
-import { getResponsiveValue, useBreakpoint } from '@vtex/admin-ui-react'
 import type { SpaceTokens, CSSPropAutocomplete } from '@vtex/admin-ui-core'
 import { cx } from '@vtex/admin-ui-core'
 
+import type { ResponsiveProp } from '../use-breakpoint'
+import { getResponsiveValue, useBreakpoint } from '../use-breakpoint'
 import { columnsStyle, columnsTheme } from './columns.css'
 
 export const Columns = forwardRef(function Columns(
