+++ packages/admin-ui/src/flex/flex.tsx
@@ -1,7 +1,7 @@
 import type { ComponentPropsWithoutRef, Ref } from 'react'
 import React, { forwardRef } from 'react'
 import type * as CSS from 'csstype'
-import type { ResponsiveProp, ResponsiveValue } from '@vtex/admin-ui-react'
+import type { ResponsiveProp, ResponsiveValue } from '../use-breakpoint'
 import { flexTheme, flexStyle } from './flex.css'
 import { cx } from '@vtex/admin-ui-core'
 import { get } from '@vtex/admin-ui-util'
