+++ packages/admin-ui/src/icons/icons.ts
@@ -1,0 +1,108 @@
+export type { IconProps } from '@vtex/phosphor-icons'
+export {
+  IconCaretUp,
+  IconCaretDown,
+  IconCaretLeft,
+  IconCaretRight,
+  IconArrowUp,
+  IconArrowDown,
+  IconArrowLeft,
+  IconArrowRight,
+  IconPlus,
+  IconMinus,
+  IconX,
+  IconChatText,
+  IconCopySimple,
+  IconHeart,
+  IconFlag,
+  IconPaperPlaneTilt,
+  IconPrinter,
+  IconCheck,
+  IconCalendarBlank,
+  IconAlarm,
+  IconClock,
+  IconArrowCounterClockwise,
+  IconArrowsClockwise,
+  IconArrowUUpLeft,
+  IconMapPin,
+  IconEye,
+  IconEyeSlash,
+  IconLockKey,
+  IconLockKeyOpen,
+  IconMagnifyingGlass,
+  IconDotsThreeVertical,
+  IconArrowLineDown,
+  IconArrowLineUp,
+  IconFaders,
+  IconTrash,
+  IconPencil,
+  IconArrowSquareOut,
+  IconArrowUpRight,
+  IconArrowCircleUpRight,
+  IconArrowSquareUpRight,
+  IconArchive,
+  IconDotsSixVertical,
+  IconSquaresFour,
+  IconRows,
+  IconChartLineUp,
+  IconCreditCard,
+  IconBarcode,
+  IconGift,
+  IconUser,
+  IconUsers,
+  IconTruck,
+  IconStorefront,
+  IconCurrencyCircleDollar,
+  IconTicket,
+  IconHandbagSimple,
+  IconRocketLaunch,
+  IconHouse,
+  IconShoppingCartSimple,
+  IconTag,
+  IconMegaphone,
+  IconLayout,
+  IconPackage,
+  IconGear,
+  IconGearSix,
+  IconTreeStructure,
+  IconCirclesThreePlus,
+  IconNotebook,
+  IconCode,
+  IconActivity,
+  IconHeadset,
+  IconIdentificationCard,
+  IconKey,
+  IconReceipt,
+  IconGlobeHemisphereWest,
+  IconSignOut,
+  IconList,
+  IconChartBar,
+  IconCloudArrowUp,
+  IconLink,
+  IconLinkBreak,
+  IconPaperclip,
+  IconEnvelope,
+  IconFileText,
+  IconImageSquare,
+  IconMagnifyingGlassPlus,
+  IconCornersOut,
+  IconListDashes,
+  IconListNumbers,
+  IconTextAlignLeft,
+  IconTextAlignCenter,
+  IconTextAlignRight,
+  IconTextBolder,
+  IconTextItalic,
+  IconTextUnderline,
+  IconFolder,
+  IconQuestion,
+  IconBell,
+  IconXCircle,
+  IconXOctagon,
+  IconWarning,
+  IconWarningCircle,
+  IconCheckCircle,
+  IconCube,
+  IconShareNetwork,
+  IconStack,
+} from '@vtex/phosphor-icons'
