+++ packages/admin-ui/src/icons/index.ts
@@ -1,0 +1,2 @@
+export * from './icon'
+export * from './icons'
