+++ packages/admin-ui/src/index.ts
@@ -1,11 +1,13 @@
 export * from '@vtex/admin-ui-core'
-export * from '@vtex/admin-ui-react'
 export * from '@vtex/admin-ui-hooks'
 export * from '@vtex/admin-ui-util'
 
 export * from './experimental'
 
+export * from './theme-provider'
 export * from './use-collapse'
+export * from './use-breakpoint'
+export * from './icons'
 export * from './center'
 export * from './inline'
 export * from './bleed'
@@ -58,113 +60,3 @@ export {
   isFirefox,
   isEdge,
 } from 'react-device-detect'
-
-export {
-  IconCaretUp,
-  IconCaretDown,
-  IconCaretLeft,
-  IconCaretRight,
-  IconArrowUp,
-  IconArrowDown,
-  IconArrowLeft,
-  IconArrowRight,
-  IconPlus,
-  IconMinus,
-  IconX,
-  IconChatText,
-  IconCopySimple,
-  IconHeart,
-  IconFlag,
-  IconPaperPlaneTilt,
-  IconPrinter,
-  IconCheck,
-  IconCalendarBlank,
-  IconAlarm,
-  IconClock,
-  IconArrowCounterClockwise,
-  IconArrowsClockwise,
-  IconArrowUUpLeft,
-  IconMapPin,
-  IconEye,
-  IconEyeSlash,
-  IconLockKey,
-  IconLockKeyOpen,
-  IconMagnifyingGlass,
-  IconDotsThreeVertical,
-  IconArrowLineDown,
-  IconArrowLineUp,
-  IconFaders,
-  IconTrash,
-  IconPencil,
-  IconArrowSquareOut,
-  IconArrowUpRight,
-  IconArrowCircleUpRight,
-  IconArrowSquareUpRight,
-  IconArchive,
-  IconDotsSixVertical,
-  IconSquaresFour,
-  IconRows,
-  IconChartLineUp,
-  IconCreditCard,
-  IconBarcode,
-  IconGift,
-  IconUser,
-  IconUsers,
-  IconTruck,
-  IconStorefront,
-  IconCurrencyCircleDollar,
-  IconTicket,
-  IconHandbagSimple,
-  IconRocketLaunch,
-  IconHouse,
-  IconShoppingCartSimple,
-  IconTag,
-  IconMegaphone,
-  IconLayout,
-  IconPackage,
-  IconGear,
-  IconGearSix,
-  IconTreeStructure,
-  IconCirclesThreePlus,
-  IconNotebook,
-  IconCode,
-  IconActivity,
-  IconHeadset,
-  IconIdentificationCard,
-  IconKey,
-  IconReceipt,
-  IconGlobeHemisphereWest,
-  IconSignOut,
-  IconList,
-  IconChartBar,
-  IconCloudArrowUp,
-  IconLink,
-  IconLinkBreak,
-  IconPaperclip,
-  IconEnvelope,
-  IconFileText,
-  IconImageSquare,
-  IconMagnifyingGlassPlus,
-  IconCornersOut,
-  IconListDashes,
-  IconListNumbers,
-  IconTextAlignLeft,
-  IconTextAlignCenter,
-  IconTextAlignRight,
-  IconTextBolder,
-  IconTextItalic,
-  IconTextUnderline,
-  IconFolder,
-  IconQuestion,
-  IconBell,
-  IconXCircle,
-  IconXOctagon,
-  IconWarning,
-  IconWarningCircle,
-  IconCheckCircle,
-  IconCube,
-  IconShareNetwork,
-  IconStack,
-} from '@vtex/phosphor-icons'
-
-export type { IconProps } from '@vtex/phosphor-icons'
