+++ packages/admin-ui/src/test-utils.tsx
@@ -3,7 +3,7 @@ import React, { useRef } from 'react'
 import { axe, toHaveNoViolations } from 'jest-axe'
 import type { RenderOptions } from '@testing-library/react'
 import { render as baseRender, fireEvent, screen } from '@testing-library/react'
-import { ThemeProvider } from '@vtex/admin-ui-react'
+import { ThemeProvider } from './theme-provider'
 import type { ReactElement } from 'react'
 import { renderHook, act } from '@testing-library/react-hooks'
 import userEvent from '@testing-library/user-event'
