+++ packages/admin-ui/src/use-media-query/index.ts
@@ -1,2 +1,1 @@
 export * from './use-media-query'
-export * from './use-breakpoint'
